#52047 Make sure the MQTT server is connected when start returns
The previous behaviour was that the connector did not reconnect if the
first connection failed. This commit changes that and there is a loop
which tries to reconnect to the MQTT server every second until it
succeeds.

It would have been good it the connector crashed if it was not
able to connect to the MQTT broker but Exceptions seems to be caught
higher up call chain...

Change-Id: Ie50af864840385bd31c404c5a8ff731f8cf4cc18